refactor: settle comments' file location
